requirements.txt
-i https://pypi.org/simple
aiofiles==24.1.0; python_version >= '3.8'
aiogram==3.22.0; python_version >= '3.9'
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.15; python_version >= '3.9'
aiosignal==1.4.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
asgiref==3.11.1; python_version >= '3.9'
async-timeout==5.0.1; python_version < '3.11'
attrs==26.1.0; python_version >= '3.9'
certifi==2026.4.22; python_version >= '3.7'
charset-normalizer==3.4.7; python_version >= '3.7'
dj-database-url==3.0.1
django==4.2.30; python_version >= '3.8'
django-auditlog==3.3.0; python_version >= '3.9'
django-axes==8.0.0; python_version >= '3.9'
django-environ==0.13.0; python_version >= '3.9' and python_version < '4'
django-ratelimit==4.1.0; python_version >= '3.7'
frozenlist==1.8.0; python_version >= '3.9'
idna==3.13; python_version >= '3.8'
magic-filter==1.0.12; python_version >= '3.7'
multidict==6.7.1; python_version >= '3.9'
phonenumbers==9.0.28; python_version >= '2.5'
propcache==0.4.1; python_version >= '3.9'
psycopg2-binary==2.9.12; python_version >= '3.9'
pydantic==2.11.10; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
requests==2.32.5; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.5; python_version >= '3.8'
tenacity==9.1.2; python_version >= '3.9'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
urllib3==2.6.3; python_version >= '3.9'
whatsapp-api-client-python==0.0.53; python_version >= '3.7'
whitenoise==6.11.0; python_version >= '3.9'
yarl==1.22.0; python_version >= '3.9'
